// pc/src/services/apis/models/StatisticCenterVo.ts
export interface StatisticCenterVo {
    /**
     * 发改-工程研究中心(产业技术创新中心)省级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count1?: number;
    /**
     * 发改-工程研究中心(产业技术创新中心)市级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count2?: number;
    /**
     * 商务-外资研发中心省级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count3?: number;
    /**
     * 商务-外资研发中心市级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count4?: number;
    /**
     * 科技-重点实验室省级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count5?: number;
    /**
     * 科技-重点实验室市级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count6?: number;
    /**
     * 科技-重点实验室国家级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count7?: number;
    /**
     * 工信-企业技术中心省级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count8?: number;
    /**
     * 工信-企业技术中心市级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count9?: number;
    /**
     * 科技-工程技术研究中心省级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count10?: number;
    /**
     * 科技-工程技术研究中心市级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count11?: number;
    /**
     * 科技-院士工作站省级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count12?: number;
    /**
     * 科技-院士工作站市级
     * @type {number}
     * @memberof StatisticCenterVo
     */
    count13?: number;
}