// pc/src/services/apis/models/StatisticInnovativeClusterVo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 企业全生命周期
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StatisticInnovativeClusterVo
 */
export interface StatisticInnovativeClusterVo {
    /**
     * 名称
     * @type {string}
     * @memberof StatisticInnovativeClusterVo
     */
    name?: string;
    /**
     * 数量
     * @type {number}
     * @memberof StatisticInnovativeClusterVo
     */
    count?: number;
    /**
     * 
     * @type {Array<StatisticInnovativeClusterVo>}
     * @memberof StatisticInnovativeClusterVo
     */
    children?: Array<StatisticInnovativeClusterVo>;
}


/**
 * Check if a given object implements the StatisticInnovativeClusterVo interface.
 */
export function instanceOfStatisticInnovativeClusterVo(value: object): value is StatisticInnovativeClusterVo {
    return true;
}

export function StatisticInnovativeClusterVoFromJSON(json: any): StatisticInnovativeClusterVo {
    return StatisticInnovativeClusterVoFromJSONTyped(json, false);
}

export function StatisticInnovativeClusterVoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StatisticInnovativeClusterVo {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'count': json['count'] == null ? undefined : json['count'],
        'children': json['children'] == null ? undefined : ((json['children'] as Array<any>).map(StatisticInnovativeClusterVoFromJSON)),
    };
}

export function StatisticInnovativeClusterVoToJSON(json: any): StatisticInnovativeClusterVo {
    return StatisticInnovativeClusterVoToJSONTyped(json, false);
}

export function StatisticInnovativeClusterVoToJSONTyped(value?: StatisticInnovativeClusterVo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'count': value['count'],
        'children': value['children'] == null ? undefined : ((value['children'] as Array<any>).map(StatisticInnovativeClusterVoToJSON)),
    };
}